Moved action sheet logic in UIViewController extension
