+++ WooCommerce/Classes/ViewRelated/Text View Screen/TextViewViewController.swift
@@ -81,26 +81,12 @@ extension TextViewViewController {
     }
 
     func presentBackNavigationActionSheet() {
-        let actionSheetMessage = NSLocalizedString("Are you sure you want to discard these changes?",
-                                                 comment: "Action sheet title in Edit Product > Edit Pricing")
-        let actionSheet = UIAlertController(title: nil, message: actionSheetMessage, preferredStyle: .actionSheet)
-        actionSheet.view.tintColor = .text
-
-        actionSheet.addDefaultActionWithTitle(ActionSheetStrings.save) { [weak self] _ in
+        presentSaveChangesActionSheet(onSave: { [weak self] in
             self?.completeEditing()
-        }
-
-        actionSheet.addDestructiveActionWithTitle(ActionSheetStrings.discard) { [weak self] _ in
+        }, onDiscard: { [weak self] in
             self?.navigationController?.popViewController(animated: true)
-        }
-
-        actionSheet.addCancelActionWithTitle(ActionSheetStrings.cancel)
-
-        let popoverController = actionSheet.popoverPresentationController
-        popoverController?.sourceView = view
-        popoverController?.sourceRect = view.bounds
-
-        present(actionSheet, animated: true)
+        }, onCancel: {
+        })
     }
 }
 
@@ -179,14 +165,4 @@ private extension TextViewViewController {
     enum Constants {
         static let textContainerInset = UIEdgeInsets(top: 16, left: 16, bottom: 16, right: 16)
     }
-
-    enum ActionSheetStrings {
-        static let save = NSLocalizedString("Save changes",
-                                            comment: "Button title in the action sheet in the generic text editor")
-        static let discard = NSLocalizedString("Discard changes",
-                                              comment: "Button title in the action sheet in the generic text editor")
-        static let cancel = NSLocalizedString("Cancel",
-                                              comment: "Dismiss the action sheet in the generic text editor")
-    }
-
 }
